:hankey: still wrong. 溢出

+++ leetcode/0201_Bitwise_AND_of_Numbers_Range/mycode.c
@@ -1,19 +1,20 @@
-int rangeBitwiseAnd(int m, int n){
+int rangeBitwiseAnd(int m, int n) {
   int result = m;
-  for(int i=m+1;i<=n;++i){
+  for (size_t i = m + 1; i <= n; ++i) {
     result &= i;
-    if(result == 0){
+    if (result == 0) {
       return 0;
     }
   }
   return result;
 }
 
-// 8264 / 8266 test cases passed.
+
+// 8265 / 8266 test cases passed.
 // Status: Runtime Error
-// Submitted: 46 minutes ago
+// Submitted: 51 minutes ago
 // Runtime Error Message:
-// Line 5: Char 22: runtime error: signed integer overflow: 2147483647 + 1 cannot be represented in type 'int' (solution.c)
+// Line 5: Char 17: runtime error: signed integer overflow: 2147483647 + 1 cannot be represented in type 'int' (solution.c)
 // Last executed input:
-// 2147483646
+// 2147483647
 // 2147483647
